finance_gpt/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
appdirs==1.4.4
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
curl_cffi==0.6.2
dataclasses-json==0.6.4
Deprecated==1.2.14
dirtyjson==1.0.8
diskcache==5.6.3
distro==1.9.0
Django==4.1.5
django-cors-headers==4.3.1
djangorestframework==3.15.1
docstring-parser==0.15
duckduckgo_search==5.2.2
exceptiongroup==1.2.0
fastapi==0.110.1
frozendict==2.4.1
frozenlist==1.4.1
fsspec==2024.3.1
google_search_results==2.4.2
greenlet==3.0.3
h11==0.14.0
html5lib==1.1
httpcore==1.0.5
httpx==0.27.0
idna==3.6
instructor==0.4.8
Jinja2==3.1.3
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.14
langchain-community==0.0.31
langchain-core==0.1.38
langchain-openai==0.1.1
langchain-text-splitters==0.0.1
langchainhub==0.1.15
langsmith==0.1.38
llama-index==0.10.27
llama-index-agent-openai==0.2.2
llama-index-cli==0.1.11
llama-index-core==0.10.27
llama-index-embeddings-openai==0.1.7
llama-index-indices-managed-llama-cloud==0.1.5
llama-index-legacy==0.9.48
llama-index-llms-openai==0.1.14
llama-index-multi-modal-llms-openai==0.1.4
llama-index-program-openai==0.1.5
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.13
llama-index-readers-llama-parse==0.1.4
llama-parse==0.4.0
llama_cpp_python==0.2.58
llamaindex-py-client==0.1.16
lxml==5.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
mdurl==0.1.2
multidict==6.0.5
multitasking==0.0.11
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.8.1
nselib==0.7
numpy==1.26.4
openai==1.16.1
orjson==3.10.0
packaging==23.2
pandas==2.2.1
peewee==3.17.1
pillow==10.3.0
pycparser==2.22
pydantic==2.6.4
pydantic-settings==2.2.1
pydantic_core==2.16.3
Pygments==2.17.2
PyMuPDF==1.24.1
PyMuPDFb==1.24.1
pypdf==4.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
rich==13.7.1
scipy==1.13.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.29
sqlparse==0.4.4
sse-starlette==2.0.0
starlette==0.37.2
starlette-context==0.3.6
striprtf==0.0.26
tavily-python==0.3.3
tenacity==8.2.3
tiktoken==0.6.0
tqdm==4.66.2
typer==0.9.4
types-requests==2.31.0.20240403
typing-inspect==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.29.0
webencodings==0.5.1
wikipedia==1.4.0
wrapt==1.16.0
yarl==1.9.4
yfinance==0.2.37
